Rename tracer test and move to parent package
