+++ opentracing-spring-cloud/src/test/java/io/opentracing/contrib/spring/cloud/TracerConfig.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.contrib.spring.cloud.web;
+package io.opentracing.contrib.spring.cloud;
 
 import io.opentracing.Tracer;
